Squashed 'lispBM/lispBM/' changes from 3b0b6d3a..f41825c1
f41825c1 fix bug in loop. classic case of not saving the environment
fa1eec03 moving prototypes of some static functions to a dedicated spot close to top of eval_cps.c
619c35a9 applying backpressure mechanism for events as suggested by Benjamin
90bb82e4 change how events are processed when the mailbox of the event handler is full. Also removed a commed about dropping a const attribute being iffy where no const attribute was dropped...
3e6b8f06 change process_events so that if event hnadler mailbox is full, the event is retried later.
66d2edf9 moved variables.lisp into beginners directory and added a zipit script for packaging of windows executable
2045fff7 update winmakefile to populate build dir with dlls
7b3e32ff adding windows library binaries
4642ef84 history file tweak for windows. History does not seem to work, at least not on wine
9eb347f7 added a variables.lisp example in repl/examples
cc7ce9bb cleaning windows repl
f34ea7f7 SDL support in windows repl
74e7ec1a Added png support on the Windows repl.
9a3f9ed4 Towards vesctcp on the windows repl
ac3427c6 towards a repl for Windows
36ca0267 towards an LBM repl for WINDOWS
d177dc5d clean out some warnings from libschrift and update repl/example/ttf_font to work again
c49090ba update to note on sharing
fedf987e Writing down notes about sharing detection from discussion with Lennart while I still somewhat remember the idea
0a934896 Merge pull request #38 from laxsjo/fix/consistent-escape-sequences
efc65d71 remove the startup-entry field from lbm_image, it was unused
5d8523ee adding comment and fixing flatten_size so that cyclic and sharing is dissallowed for now
b8a76e98 Towards more correct sharing and cycle detection in preparation for sharing restoration upon unflattening
aa486c57 fix: Make string and character escapes consistent

git-subtree-dir: lispBM/lispBM
git-subtree-split: f41825c1bbf8c276cc1b9aecd8dc37b835f6ff75

+++ src/extensions/ttf_extensions.c
@@ -675,7 +675,7 @@ lbm_value ttf_text_bin(lbm_value *args, lbm_uint argn) {
       }
       x_n += x_shift;
       y_n += y_shift;
-      y_n += y_offset;
+      y_n += (float)y_offset;
 
       image_buffer_t src;
       src.width = (uint16_t)width;
